CODE/GATHER_DATA/requirements.txt
datetime
pandas
pyarrow
tefas-crawler == 0.3.4
